Completed the Paragraph Styles. Kerning/Tracking is now a percentage of the Fontsize. Accepted Range is -300 % to +300%
git-svn-id: svn://scribus.net/branches/Version13x/Scribus@2382 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/mpalette.h
@@ -286,7 +286,7 @@ public slots:
 	void setLspMode(int id);
 	void setLsp(double r);
 	void setSize(int s);
-	void setExtra(double e);
+	void setExtra(int e);
 	void setDvals(double left, double top, double bottom, double right);
 	void ChangeScaling();
 	void setLvalue(double scx, double scy, double x, double y);
